Modularize extensions, integrate object detection model
